requirements.txt
amqp==2.5.2
appdirs==1.4.3
asgiref
asn1crypto==0.24.0
astor==0.8.1
attrs==21.4.0
autoflake==1.4
autopep8==1.4.3
Babel==2.6.0
beautifulsoup4==4.7.1
billiard==3.6.1.0
black==22.12.0
celery==4.4.0
certifi==2018.11.29
cffi==1.15.0
chardet==3.0.4
click==8.0.4
coverage==5.3
cryptography==2.5
deepdiff==5.0.2
Dickens==2.0.0
Django==3.2.12
django-admin==1.3.2
django-bulk-update==2.2.0
django-cachalot==2.5.1
django-debug-toolbar==3.2.4
django-excel-response2==2.0.8
django-extensions==3.1.5
django-filter==21.1
django-lint==2.0.4
django-redis-cache==2.1.0
django-six==1.0.4
django-static-jquery==2.1.4
djangorestframework==3.13.1
flake8==3.9.2
flower==0.9.2
flynt==0.64
future==0.18.2
gitdb==4.0.2
GitPython==3.1.0
graphviz==0.10.1
gunicorn==19.9.0
httplib2==0.12.0
idna==2.8
importlab==0.7
importlib-metadata==1.4.0
isort==5.8.0
kombu==4.6.7
logilab-astng==0.24.3
logilab-common==1.4.2
Markdown==3.0.1
mccabe==0.6.1
more-itertools==8.1.0
networkx
ninja==1.10.2.3
ordered-set
packaging==19.0
pendulum==2.0.5
pep517==0.12.0
pep8==1.7.1
pip-review==1.0
pip-tools
pipdeptree==2.2.1
pre-commit==3.3.2
psycopg2-binary==2.8
pycodestyle==2.7.0
pycparser==2.19
pydot==1.4.1
pyeda==0.28.0
pyflakes==2.3.1
pykerberos==1.2.1
pylint==0.28.0
pyparsing==2.3.1
python-dateutil==2.8.1
python-json-logger==2.0.7
python-memcached==1.59
pytype==2021.3.22
pytz==2018.9
pytzdata==2019.3
PyYAML==6.0
redis==3.4.1
requests==2.21.0
requests-kerberos==0.12.0
ruff
screen==1.0.1
shell==1.0.1
six==1.12.0
smmap==3.0.1
soupsieve==1.7.3
SQLAlchemy==1.2.18
sqlparse==0.2.4
toml==0.10.2
tomli
tornado==5.1.1
treelib==1.6.1
typed-ast==1.5.2
unify==0.5
urllib3==1.24.1
vine==1.3.0
xlwt==1.3.0
XStatic-Patternfly-Bootstrap-Treeview==2.1.3.2
yamale==1.8.1
yapf==0.31.0
zipp==2.0.0
